requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.6
langchain==0.3.23
langchain-core==0.3.54
langchain-community>=0.3.0
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
numpy>=1.26.3
pandas>=2.2.0
jinja2>=3.1.3
Pillow>=10.0.0